requirements.txt
python-dotenv
langchain
openai
slack-sdk
slack-bolt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
aiocache
aioredis
requests
types-requests

fastapi
slack_bolt
aiohttp
httpx
# needed for voice
soundfile
keyboard
elevenlabs
uvicorn[standard]
# ASGI for azure web  service
uvloop
httptools
gunicorn

# tests
pytest
pytest-cov
starlette
pinecone-client
